refactor: add ESLint config and fix all linting errors (#95)
* refactor: eslint config and fixs errors with new config

* feat: eslint rules

* chore: fix code style issues to comply with new ESLint config

+++ libs/s3/src/s3.module.ts
@@ -1,9 +1,11 @@
-import { Inject, Module, OnApplicationShutdown } from '@nestjs/common';
-import type { S3ModuleOptions } from './interfaces';
-import { S3Service } from './s3.service';
-import { ConfigurableModuleClass, MODULE_OPTIONS_TOKEN } from './s3.module-definition';
-import { S3_CLIENT } from './constants';
 import { S3Client } from '@aws-sdk/client-s3';
+import { Inject, Module, OnApplicationShutdown } from '@nestjs/common';
+
+import { S3_CLIENT } from './constants';
+import { ConfigurableModuleClass, MODULE_OPTIONS_TOKEN } from './s3.module-definition';
+import { S3Service } from './s3.service';
+
+import type { S3ModuleOptions } from './interfaces';
 
 @Module({
     providers: [
@@ -28,7 +30,9 @@ export class S3Module extends ConfigurableModuleClass implements OnApplicationSh
         super();
     }
 
-    async onApplicationShutdown() {
-        this.client.destroy();
+    onApplicationShutdown() {
+        if (this.client.destroy) {
+            this.client.destroy();
+        }
     }
 }
